requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.3
django-filter==21.1
django-graphql-jwt==0.3.4
docx2pdf==0.1.8
Faker==12.0.1
graphene==3.0
graphene-django==3.0.0b7
graphene-file-upload==1.3.0
graphql-core==3.1.7
graphql-relay==3.1.5
kombu==5.2.4
mixer==7.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.2
promise==2.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
